Moved q policies into their own file
